pyproject.toml
[project]
name = "app-store-reviews-viewer-backend"
version = "0.1.0"
description = "Recent iOS AppStore Reviews Viewer Backend"
authors = [{ name = "Mikhail Vybornyi", email = "vbrn.mv@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-httpx>=0.35.0",
    "pytest-mock>=3.14.1",
    "autoflake>=2.3.1",
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = [
    "--failed-first",
    "--strict-markers",
    "--show-capture=no",
    "--disable-warnings",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = '3.13'
ignore_missing_imports = true
strict = false
show_error_codes = true

allow_redefinition = true
allow_redefinition_new = true
local_partial_types = true

disable_error_code = [
    "assignment",
    "misc",
    "override",
    "no-any-return",
    "var-annotated", # allows: foo = [], bar = {}
]

[tool.autoflake]
ignore-init-module-imports = true
remove-all-unused-imports = true
remove-unused-variables = true
in-place = true
expand-star-imports = true
